services/image-processor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
asyncpg==0.29.0
redis==5.0.1
aiohttp==3.9.1
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.25.2
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0